[CORE] Move trait GlutenPlan to gluten-core
