+++ backends-velox/src/main/scala/org/apache/spark/sql/execution/BaseArrowScanExec.scala
@@ -17,11 +17,13 @@
 package org.apache.spark.sql.execution
 
 import org.apache.gluten.columnarbatch.ArrowBatches
-import org.apache.gluten.extension.GlutenPlan
+import org.apache.gluten.execution.GlutenPlan
 import org.apache.gluten.extension.columnar.transition.Convention
 
 trait BaseArrowScanExec extends GlutenPlan {
   final override def batchType(): Convention.BatchType = {
     ArrowBatches.ArrowJavaBatch
   }
+
+  final override def rowType0(): Convention.RowType = Convention.RowType.None
 }
