+++ gluten-core/src/main/scala/org/apache/gluten/execution/ColumnarToColumnarExec.scala
@@ -32,7 +32,7 @@ abstract class ColumnarToColumnarExec(from: Convention.BatchType, to: Convention
   extends ColumnarToColumnarTransition
   with Convention.KnownBatchType
   with Convention.KnownRowTypeForSpark33AndLater
-  with ConventionReq.KnownChildrenConventions {
+  with ConventionReq.KnownChildConvention {
 
   def child: SparkPlan
   protected def mapIterator(in: Iterator[ColumnarBatch]): Iterator[ColumnarBatch]
@@ -60,7 +60,7 @@ abstract class ColumnarToColumnarExec(from: Convention.BatchType, to: Convention
     Convention.RowType.None
   }
 
-  override def requiredChildrenConventions(): Seq[ConventionReq] = List(
+  override def requiredChildConvention(): Seq[ConventionReq] = List(
     ConventionReq.of(ConventionReq.RowType.Any, ConventionReq.BatchType.Is(from)))
 
   override protected def doExecute(): RDD[InternalRow] = throw new UnsupportedOperationException()
