+++ gluten-substrait/src/main/scala/org/apache/gluten/execution/ColumnarToRowExecBase.scala
@@ -17,9 +17,8 @@
 package org.apache.gluten.execution
 
 import org.apache.gluten.backendsapi.BackendsApiManager
-import org.apache.gluten.extension.GlutenPlan
 import org.apache.gluten.extension.columnar.transition.{Convention, ConventionReq}
-import org.apache.gluten.extension.columnar.transition.ConventionReq.KnownChildrenConventions
+import org.apache.gluten.extension.columnar.transition.ConventionReq.KnownChildConvention
 
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.rdd.RDD
@@ -31,7 +30,8 @@ import org.apache.spark.sql.execution.{ColumnarToRowTransition, SparkPlan}
 abstract class ColumnarToRowExecBase(child: SparkPlan)
   extends ColumnarToRowTransition
   with GlutenPlan
-  with KnownChildrenConventions {
+  with KnownChildConvention
+  with ValidatablePlan {
 
   // Note: "metrics" is made transient to avoid sending driver-side metrics to tasks.
   @transient override lazy val metrics =
@@ -58,7 +58,7 @@ abstract class ColumnarToRowExecBase(child: SparkPlan)
     doExecuteInternal()
   }
 
-  override def requiredChildrenConventions(): Seq[ConventionReq] = {
+  override def requiredChildConvention(): Seq[ConventionReq] = {
     List(
       ConventionReq.ofBatch(
         ConventionReq.BatchType.Is(BackendsApiManager.getSettings.primaryBatchType)))
