+++ gluten-substrait/src/main/scala/org/apache/gluten/extension/columnar/validator/Validators.scala
@@ -19,7 +19,7 @@ package org.apache.gluten.extension.columnar.validator
 import org.apache.gluten.GlutenConfig
 import org.apache.gluten.backendsapi.{BackendsApiManager, BackendSettingsApi}
 import org.apache.gluten.exception.GlutenNotSupportException
-import org.apache.gluten.execution.{BasicScanExecTransformer, ColumnarCoalesceExec, ColumnarUnionExec, ExpandExecTransformer, HashAggregateExecBaseTransformer, LimitExecTransformer, ProjectExecTransformer, ScanTransformerFactory, SortExecTransformer, TakeOrderedAndProjectExecTransformer, WindowExecTransformer, WindowGroupLimitExecTransformer, WriteFilesExecTransformer}
+import org.apache.gluten.execution._
 import org.apache.gluten.expression.ExpressionUtils
 import org.apache.gluten.extension.columnar.FallbackTags
 import org.apache.gluten.extension.columnar.offload.OffloadJoin
@@ -268,8 +268,7 @@ object Validators {
         val transformer = HashAggregateExecBaseTransformer.from(plan)
         transformer.doValidate().toValidatorOutcome()
       case plan: UnionExec =>
-        val transformer = ColumnarUnionExec(plan.children)
-        transformer.doValidate().toValidatorOutcome()
+        pass()
       case plan: ExpandExec =>
         val transformer = ExpandExecTransformer(plan.projections, plan.output, plan.child)
         transformer.doValidate().toValidatorOutcome()
@@ -302,8 +301,7 @@ object Validators {
             plan.isSkewJoin)
         transformer.doValidate().toValidatorOutcome()
       case plan: BroadcastExchangeExec =>
-        val transformer = ColumnarBroadcastExchangeExec(plan.mode, plan.child)
-        transformer.doValidate().toValidatorOutcome()
+        pass()
       case bhj: BroadcastHashJoinExec =>
         val transformer = BackendsApiManager.getSparkPlanExecApiInstance
           .genBroadcastHashJoinExecTransformer(
@@ -361,9 +359,7 @@ object Validators {
         )
         transformer.doValidate().toValidatorOutcome()
       case plan: CoalesceExec =>
-        ColumnarCoalesceExec(plan.numPartitions, plan.child)
-          .doValidate()
-          .toValidatorOutcome()
+        pass()
       case plan: GlobalLimitExec =>
         val (limit, offset) =
           SparkShimLoader.getSparkShims.getLimitAndOffsetFromGlobalLimit(plan)
