+++ gluten-substrait/src/main/scala/org/apache/spark/sql/execution/ColumnarWriteFilesExec.scala
@@ -18,9 +18,9 @@ package org.apache.spark.sql.execution
 
 import org.apache.gluten.backendsapi.BackendsApiManager
 import org.apache.gluten.exception.GlutenException
-import org.apache.gluten.extension.GlutenPlan
+import org.apache.gluten.execution.GlutenPlan
+import org.apache.gluten.extension.columnar.transition.{Convention, ConventionReq}
 import org.apache.gluten.extension.columnar.transition.Convention.RowType
-import org.apache.gluten.extension.columnar.transition.ConventionReq
 import org.apache.gluten.sql.shims.SparkShimLoader
 
 import org.apache.spark.TaskContext
@@ -44,14 +44,13 @@ abstract class ColumnarWriteFilesExec protected (
     override val right: SparkPlan)
   extends BinaryExecNode
   with GlutenPlan
-  with ConventionReq.KnownChildrenConventions
   with ColumnarWriteFilesExec.ExecuteWriteCompatible {
 
   val child: SparkPlan = left
 
   override lazy val references: AttributeSet = AttributeSet.empty
 
-  override def requiredChildrenConventions(): Seq[ConventionReq] = {
+  override def requiredChildConvention(): Seq[ConventionReq] = {
     List(
       ConventionReq.ofBatch(
         ConventionReq.BatchType.Is(BackendsApiManager.getSettings.primaryBatchType)))
@@ -67,6 +66,7 @@ abstract class ColumnarWriteFilesExec protected (
    *
    * Since https://github.com/apache/incubator-gluten/pull/6745.
    */
+  override def batchType(): Convention.BatchType = BackendsApiManager.getSettings.primaryBatchType
   override def rowType0(): RowType = {
     RowType.VanillaRow
   }
