+++ gluten-ut/spark34/src/test/scala/org/apache/spark/sql/execution/FallbackStrategiesSuite.scala
@@ -17,13 +17,14 @@
 package org.apache.spark.sql.execution
 
 import org.apache.gluten.GlutenConfig
-import org.apache.gluten.execution.BasicScanExecTransformer
-import org.apache.gluten.extension.{GlutenPlan, GlutenSessionExtensions}
+import org.apache.gluten.backendsapi.BackendsApiManager
+import org.apache.gluten.execution.{BasicScanExecTransformer, GlutenPlan}
+import org.apache.gluten.extension.GlutenSessionExtensions
 import org.apache.gluten.extension.columnar.{FallbackTags, RemoveFallbackTagRule}
 import org.apache.gluten.extension.columnar.ColumnarRuleApplier.ColumnarRuleCall
 import org.apache.gluten.extension.columnar.MiscColumnarRules.RemoveTopmostColumnarToRow
 import org.apache.gluten.extension.columnar.heuristic.{ExpandFallbackPolicy, HeuristicApplier}
-import org.apache.gluten.extension.columnar.transition.InsertBackendTransitions
+import org.apache.gluten.extension.columnar.transition.{Convention, InsertBackendTransitions}
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{GlutenSQLTestsTrait, SparkSession}
@@ -229,6 +230,8 @@ private object FallbackStrategiesSuite {
 
   // For replacing LeafOp.
   case class LeafOpTransformer() extends LeafExecNode with GlutenPlan {
+    override def batchType(): Convention.BatchType = BackendsApiManager.getSettings.primaryBatchType
+    override def rowType0(): Convention.RowType = Convention.RowType.None
     override protected def doExecute(): RDD[InternalRow] = throw new UnsupportedOperationException()
     override def output: Seq[Attribute] = Seq.empty
   }
@@ -237,6 +240,8 @@ private object FallbackStrategiesSuite {
   case class UnaryOp1Transformer(override val child: SparkPlan)
     extends UnaryExecNode
     with GlutenPlan {
+    override def batchType(): Convention.BatchType = BackendsApiManager.getSettings.primaryBatchType
+    override def rowType0(): Convention.RowType = Convention.RowType.None
     override protected def doExecute(): RDD[InternalRow] = throw new UnsupportedOperationException()
     override def output: Seq[Attribute] = child.output
     override protected def withNewChildInternal(newChild: SparkPlan): UnaryOp1Transformer =
